// frontend/src/api/generated/model/lostFoundItemKind.ts
export type LostFoundItemKind = typeof LostFoundItemKind[keyof typeof LostFoundItemKind];


export const LostFoundItemKind = {
  found: 'found',
  lost: 'lost',
} as const;